requirements.txt
# ZeddiHub Tools Desktop v1.0.0
# GUI Dependencies
customtkinter>=5.2.0
Pillow>=10.0.0
cryptography>=41.0.0

# Network
requests>=2.31.0

# Standard library (included with Python):
# tkinter, json, os, sys, socket, struct, re, time, shutil, threading, urllib
